requirements.txt
torch>=2.0
numpy>=1.23
Pillow>=9.0
